criei arquivo de utils para separar funções
